refactor: restructure claim path pointer module and update imports
