+++ crates/cloud-wallet-openid4vc/src/core/claim_path_pointer.rs
