+++ crates/cloud-wallet-openid4vc/src/core/mod.rs
@@ -1,0 +1,1 @@
+pub mod claim_path_pointer;
