+++ crates/cloud-wallet-openid4vc/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod core;
 pub mod errors;
 pub mod oid4vci;
 pub mod utils;
